add diet endpoints

+++ src/main/java/br/unifil/macrostruct/repository/FoodRepository.java
@@ -3,5 +3,5 @@ package br.unifil.macrostruct.repository;
 import br.unifil.macrostruct.model.FoodEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FoodRepository extends JpaRepository<FoodEntity, Long> {
+public interface FoodRepository extends JpaRepository<FoodEntity, Integer> {
 }
